[FEATURE] ott 상세 조회 에러 페이지 연동 closed #15
