+++ src/main/java/com/boa/conlog/ott/controller/OttController.java
@@ -1,5 +1,6 @@
 package com.boa.conlog.ott.controller;
 
+import com.boa.conlog.ott.exception.OttNotFoundException;
 import com.boa.conlog.ott.model.dto.OttDTO;
 import com.boa.conlog.ott.model.service.OttService;
 import org.springframework.stereotype.Controller;
@@ -32,9 +33,13 @@ public class OttController {
 
     /* OTT 상세 조회 */
     @GetMapping("/one/{ottNo}")
-    public String selectOneOtt(@PathVariable int ottNo, Model model){
+    public String selectOneOtt(@PathVariable int ottNo, Model model) throws OttNotFoundException {
 
         OttDTO ott = ottService.selectOneOtt(ottNo);
+
+        if(ott == null){
+            throw new OttNotFoundException("해당 번호의 OTT는 찾을 수 없습니다.");
+        }
 
         model.addAttribute("oneOtt", ott);
