+++ src/main/java/com/boa/conlog/ott/exception/OttNotFoundException.java
@@ -1,0 +1,9 @@
+package com.boa.conlog.ott.exception;
+
+public class OttNotFoundException extends Exception{
+
+    public OttNotFoundException(String msg){
+
+        super(msg);
+    }
+}
