+++ src/main/java/com/boa/conlog/ott/exception/OttRegistException.java
@@ -1,0 +1,8 @@
+package com.boa.conlog.ott.exception;
+
+public class OttRegistException extends Exception{
+
+    public OttRegistException(String msg){
+        super(msg);
+    }
+}
