+++ src/main/resources/templates/ott/error/notFound.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+</head>
+<body>
+    <div th:replace="common/header.html"></div>
+
+    <h1 th:text="${ exception.message }" align="center"></h1>
+</body>
+</html>
