+++ src/main/resources/templates/ott/list.html
@@ -9,6 +9,7 @@
             alert(successMessage)
         }
     </script>
+    <link rel="stylesheet" th:href="@{/css/review.css}"/>
 </head>
 <body>
     <div th:replace="common/header.html"></div>
