+++ src/main/resources/templates/ott/oneOtt.html
@@ -10,6 +10,7 @@
             alert(successMessage);
         }
     </script>
+    <link rel="stylesheet" th:href="@{/css/review.css}"/>
 </head>
 <body>
     <div th:replace="common/header.html"></div>
